+++ src/groups/index.tsx
@@ -5,6 +5,7 @@ export { default as Arrow } from "./Arrow";
 export { default as Camera } from "./Camera";
 export { default as BoundingBox } from "./BoundingBox";
 export { default as Capsule } from "./Capsule";
+export { default as FBX } from "./FBX";
 export { default as GLTF } from "./GLTF";
 export { default as Group, generateGroupProps, GroupPropTypes } from "./Group";
 export { default as OBJ } from "./OBJ";