+++ CHANGELOG.md
@@ -2,10 +2,19 @@
 
 # Unreleased
 
-# 1.0.2
+# 1.0.5
 
-Cleaned up readme and fix image links
+- 🎉 Added `FBX` which accepts the following props
+  - `fbxPath`: string
+  - `fbxURL`: string
+  - `actionIndex`: number
+  - all `Group` props
+- 📖 Added `FBX` Story
+
+# 1.0.2 - 1.0.4
+
+- 🧹 Cleaned up readme and fix image links
 
 # 1.0.0
 
-Released this library as Open Source
+- 🎊 Released this library as Open Source